+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where previously existing blocks could be lost when new blocks had validation errors, if `autosaveDrafts` was disabled
+
 ## 5.2.12 - 2024-10-21
 
 ### Added